lazy: improve and test `__repr__`
Summary:
It was [noted in #1781][3] that the `__repr__` of `LazyModule`s deserves
a modicum of special attention. This commit tests that, and improves the
`__repr__` for already-loaded lazy modules so that it’s clear that they
are still going through the `LazyModule` proxy. (Should there be a bug
in the proxy interface—say, someone expects `__setattr__` to also pass
through to the underlying module—this would make it much easier to track
down.)

[3]: https://github.com/tensorflow/tensorboard/pull/1781#pullrequestreview-197349447

Test Plan:
`bazel test //tensorboard:lazy_test` passes.

wchargin-branch: lazy-repr